upload_27/03/2023

+++ src/tasks/tasks.service.ts
@@ -7,75 +7,21 @@ import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskRepository } from './tasks.repository';
 import {InjectRepository} from '@nestjs/typeorm'
 import { Task } from './task.entity';
+import { User } from 'src/auth/use.entity';
 
 @Injectable()
 export class TasksService {
   constructor(
     @InjectRepository(TaskRepository)
     private taskRepository: TaskRepository,
-  ){
+  ){}
+
+  async getTask(filterDto: GetTasksFilterDto, user: User):Promise<Task[]>{
+    return this.taskRepository.getTasks(filterDto, user)
   }
-  // private tasks: Task[] = [];
 
-  // getAllTasks(): Task[] {
-  //   return this.tasks;
-  // }
-
-  // getTaskWithFilter(filterDto: GetTasksFilterDto): Task[]{
-  //   const {status, search} = filterDto;
-  //   let tasks = this.getAllTasks();
-  //   if (status){
-  //     tasks = tasks.filter(task => task.status === status);
-  //   }
-  //   if (search){
-  //     console.log(search)
-  //     tasks = tasks.filter(task =>
-  //       task.title.includes(search)||
-  //       task.description.includes(search),
-  //     );
-  //   }
-
-
-
-  //   return tasks
-
-
-  // }
-
-  // createTask(createTaskDto: CreateTaskDto): Task {
-  //   const {title, description} = createTaskDto;
-
-  //   const task: Task = {
-  //     id: uuidv1(),
-  //     title: title,
-  //     description: description,
-  //     status: TaskStatus.OPEN
-  //   };
-  //   this.tasks.push(task)
-  //   return task;
-  // }
-  // promise_function(){
-  //   const currentPromise = new Promise((resolve, reject) => {
-  //     let condition = true;
-  //     if (condition){
-  //       setTimeout(()=>{
-  //         resolve('Success');
-  //       }, 3000);
-  //     } else{
-  //       reject('Error');
-  //     }
-  //   });
-  //   currentPromise
-  //   .then((data)=> {
-  //     console.log(data);
-  //   })
-  //   .catch(err => {
-  //     console.log(err);
-  //   })
-  // }
-
-  async getTaskById(id: number): Promise<Task>{
-    const found = await this.taskRepository.findOne({ where: { id } });
+  async getTaskById(id: number, user: User): Promise<Task>{
+    const found = await this.taskRepository.findOne({where:{id, userId : user.id}});
     
     if (!found){
       throw new NotFoundException(`Task with ID "${id} not found`)
@@ -84,12 +30,12 @@ export class TasksService {
   }
 
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task>{
-    return this.taskRepository.createTask(createTaskDto)
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task>{
+    return this.taskRepository.createTask(createTaskDto, user)
   }
 
-  async deleteTask(id: number) : Promise<any> {
-    const result =  await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User) : Promise<any> {
+    const result =  await this.taskRepository.delete({id, userId :user.id});
     if (result.affected==0){
       throw new NotFoundException(`Task with ID "${id}" not found`);
     }
@@ -99,23 +45,12 @@ export class TasksService {
   }
 
 
-  // getTaskById(id: string): Task{
-  //   const found =  this.tasks.find(task => task.id === id);
-  //   if (!found){
-  //     throw new NotFoundException(`Task with ID "${id} not found`)
-  //   }
-  //   return found
-  // }
+  async updateTaskById(id: number, status: TaskStatus, user: User){
+    const data = await this.getTaskById(id, user);
+    data.status = status;
+    data.save();
 
-  // deleteTaskById(id: string){
-  //   const found =  this.getTaskById(id);
-  //   this.tasks = this.tasks.filter(task => task.id !== id);
-  // }
-
-  // updateTaskById(id: string, status: TaskStatus){
-  //   const data = this.getTaskById(id);
-  //   data.status = status;
-  //   return data
-  // }
+    return data
+  }
 
 }
